+++ src/lightning/pytorch/loggers/logger.py
@@ -15,11 +15,11 @@
 
 import functools
 import operator
+import statistics
 from abc import ABC
 from collections import defaultdict
 from typing import Any, Callable, Dict, Mapping, Optional, Sequence
 
-import numpy as np
 from typing_extensions import override
 
 from lightning.fabric.loggers import Logger as FabricLogger
@@ -100,7 +100,7 @@ class DummyLogger(Logger):
 def merge_dicts(  # pragma: no cover
     dicts: Sequence[Mapping],
     agg_key_funcs: Optional[Mapping] = None,
-    default_func: Callable[[Sequence[float]], float] = np.mean,
+    default_func: Callable[[Sequence[float]], float] = statistics.mean,
 ) -> Dict:
     """Merge a sequence with dictionaries into one dictionary by aggregating the same keys with some given function.
 
@@ -126,7 +126,7 @@ def merge_dicts(  # pragma: no cover
         >>> d2 = {'a': 1.1, 'b': 2.2, 'v': 1, 'd': {'d1': 2, 'd2': 3}}
         >>> d3 = {'a': 1.1, 'v': 2.3, 'd': {'d3': 3, 'd4': {'d5': 1}}}
         >>> dflt_func = min
-        >>> agg_funcs = {'a': np.mean, 'v': max, 'd': {'d1': sum}}
+        >>> agg_funcs = {'a': statistics.mean, 'v': max, 'd': {'d1': sum}}
         >>> pprint.pprint(merge_dicts([d1, d2, d3], agg_funcs, dflt_func))
         {'a': 1.3,
          'b': 2.0,